+++ src/Fields/Field.php
@@ -33,6 +33,16 @@ abstract class Field extends Fluent
 
 	abstract public function getInput();
 
+	public function getHelpText()
+	{
+		if ($this->help_text) {
+			return (object) [
+				'title' => $this->help_text_title,
+				'content' => $this->help_text,
+			];
+		}
+	}
+
 	public function getLabel()
 	{
 		return Form::label($this->get('name'), $this->get('label'));
@@ -53,6 +63,7 @@ abstract class Field extends Fluent
 	public function render($errors)
 	{
 		return View::make('admin::fields.field')->with([
+			'helpText' => $this->getHelpText(),
 			'label' => $this->getLabel(),
 			'input' => $this->getInput(),
 			'errors' => $this->getErrors($errors),